pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Game Genre Analysis"
version = "0.1.0"
description = "Small app for genre analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "SQLAlchemy",
  "httpx",
  "python-dotenv",
  "alembic",
  "prefect"
]

[tool.setuptools.packages.find]
where = ["src"]